requirements.txt
mpmath==1.2.1
numpy==1.22.3
pylab-sdk==1.3.2
sympy==1.9
